make: integrate regression tests
Use `-j 1` in CI for unmangled output.

In the future this could be improved by integrating with Automake's
test support to get better output with multiple jobs. Considering
the layour of the crash tests will likely change and that currently
the tests run in less 15 seconds even with just a single thread,
this should be good enough for now.