fuzz: add OSS-Fuzz build target
OSS-Fuzz requires linking with it's C++ clang++ specifically, due to
using sanitizer link options failing with C clang. While it would be
possible to more generally allow setting a custom linker for building
the fuzz program, this wouldn't be much cleaner and would
likely introduce all sort of problems due to the (un)intended
interactions with default linkflags and libtool. And if we only want it
to be good enough for OSS-Fuzzz, we might as well just make it actually
OSS-Fuzz specific.

This will also allow us to push many changes and fixes to the
fuzzer without having to go through OSS-Fuzz repository and its
Google-account-bound CLA.