render: improve storage res fallback when PAR is set
When storage size is not set, we currently use PlayRes
as a fallback hoping it matches the original storage res.
We know this is flawed, but even if PlayRes relates to a
dimension of the source video, in case of anamorphic video
either the native display width or storage width may have
been used. We always want the latter, so if PAR information
was supplied, try to ensure we get something in the storage
aspect ratio.

Note, that this could actually also worsen the rendering for
unknown storage res cases eg if PlayRes already was the original
storage size but the subs got repurposed for a non-anamorphic
(re)encode. Though, such reused subs won't render correctly in VSFilter
too anyway, so we probably don't need to worry to much about this.