Rename orig_{width,height} and font_scale_x
font_scale_x is a PAR factor and applies to more than just font glyphs,
so par_scale_x is a more accurate and natural name.

orig_{width,height} can easily be confused with the video's
original (storage) resolution, when in fact it is the rendering
resolution (after any anamorphic desqueezing and other scaling etc)
of the video content area of the whole frame.
Thus rename to frame_content_*.