api: add fields for v4++
But no implementation yet. Implementations for the new Margin properties
and v4++'s \kt tag will be added in follow-up commits.
For RelativeTo no implementation is immediately planned, but given
xy-VSFilter also dropped support for RelativeTo, this doesn't seem
pressing. A field is already added and initalised to VSFilter's
default anyway, to spare us another ABI break if a guliverkli(2)-like
implementation is ever added.

The top margin keeps the MarginV name to not break public API
as the Event and Style properties are exposed in ass_types.h.
If we want to preserve full API compatibility even with
purely API-created ASS_Styles, we'll need to put version conditionals
into the rendering, not only the parsing, part.

If we only cared about preserving API for _reading_ the fields:
An alternative would be to alias MarginV to MarginT via anonymous union,
to preserve API while only using the clearer T/B names internally.
However, anonymous unions are not part of ISO C99, and while requiring
ISO C11 or the widespread earlier compiler extension enabling this
might be fine for *building* libass, requiring as part of our public API
seems more problematic so this just stick with the MarginV name only.

(An anonymous union would also help somewhat with easing a
 potential future API-transition from MarginV to MarginT by allowing
 API-users to already use the new name now.)

Even with API preserved, this still breaks ABI since arrays of
styles and events are exposed in the API and their element size
changed by adding new fields to those structs.