Support SSA's AlphaLevel style field
Procesing of the actual style field emulates VSFilter's behaviour
by parsing the inputs the same way and always using 0x80 for shadows.
For user supplied style overrides the field is accepted for all
script versions and the provided alpha value is applied to all colours,
continuing the existing practice of override tags not caring for
version-dependent quirks and limitations.

Fixes: https://github.com/libass/libass/issues/40