Parse ScriptType header
VSFilter uses it for Dialogue lines and Style lines
not preceeded by a styles section header.
libass requires section headers to parse styles,
but this still helps for files without a Styles
section (placed before the Events section) and
also makes our legacy-FFmpeg detection more accurate.

In fact, VSFilter takes _only_ the ScriptType
header into account for Dialogue lines and keeps
a separate version for styles which is also affected
by style section headers.
This dual versioning is not implemented by this commit and the
lack of any issue reports about it suggests such files may not
exist in practice. If however it turned out they do, an additional
version variable can be added to parser_priv and we need to decide
which one to present to API users.