Recognise v4++ files
This does not actually implement any v4++ features yet,
but prepares for commits which do.

As far as I can tell, no tool ever wrote or parsed Format lines
for v4++ files, so in the interest of better compatibility, libass'
custom format line extension is disabled when parsing v4++.
With no canonical Format line to go by, this means the internal
pseudo-format line and the new field names used for parsing v4++
are a fabrication of this commit. Albeit, the naming is corroborated
by both VSFilter's internal names, existing Margin* naming and the
naming used in asa's AS5 wiki page[1].
Notably, the new field names will leak to custom format line users
who will now be able to use them outside of v4++, but this seems mostly
harmless as we won't need to change the fabricated field names.

[1]: https://web.archive.org/web/20071030224805/http://asa.diac24.net/AS5