refactor: prefix all internal API with ass_
If static libass is linked into a binary defining functions
of the same name there will be issues. To avoid this use an
ass_ prefix for namespacing.
Before this commit we already did this for most but not yet
all internal API.

read_file is renamed to ass_load_file as ass_read_file
already exists as a public API function. All other functions
are simply prefixed with ass_.

Fixes: https://github.com/libass/libass/issues/222
Fixes: https://github.com/libass/libass/issues/654