Rough implementation of Banner/Scroll effects' fadeawaywidth/height
This should work and match VSFilter when rendering to target
display resolution equal to layout (video storage) reslution,
except that our visual quality is better due to its limited
value range (0-64 for both base bitmap and gradient bitmap before
they're multiplied together) and more abundant rounding error.

When rendering to a different target display resolution, the location
of our fade bands as implemented in this commit will not exactly match
DirectVobSub's, and the greater the scaling factor, the bigger the difference
will be. XySubFilter also exhibits this problem, although our locations don't
exactly match XySubFilter's, either, except when PlayRes == LayoutRes.

The gradient is implemented as multiple ASS_Images to avoid spending time
and memory on a bitmap multiplication on every frame, as the event keeps
scrolling and the multiplied bitmap would be uncacheable. However, this
does increase the workload of the consumer and may also harm applications
that use ASS_Images to estimate event bounding boxes.