Make Banner/Scroll fadeaway stay firmly in place when rescaling
Compared to before, this shifts the gradient by half a pixel up/left when
rendering at PlayRes, but this new position is exactly reproduced when
rendering at any other resolution.

The linear gradient now covers exactly the area from the left edge
of the video to the right edge of the fadeawaywidthth PlayRes-pixel
(exactly fadeawaywidth PlayRes-pixels in total), similarly on the right,
and similarly from the top edge of the y0th PlayRes-pixel for exactly
fadeawayheight PlayRes-pixels down, and from the bottom edge of y1st
PlayRes-pixel for exactly fadeawayheight PlayRes-pixels up. I think
this is the most intuitive arrangement.

The gradient is sampled at the centre of each display-resolution pixel.

This does NOT match any VSFilter variant in any configuration. However, for a
fixed video viewed on ever-denser displays, I think MPC-HC/BE ISR's gradient
(and for PlayRes == video res, also XySubFilter's) tends to this same result
in its limit as the display resolution grows to infinity... or rather, would
tend if you could ignore its rounding error.