Release 0.17.0
This release’s major feature is the addition of a new ASS ScriptInfo
header pair: LayoutResX and LayoutResY. The introduction has been
coordinated with active VSFilters and will soon be available in
xy-VSFilter and XySubFilter as well.
Since the beginning, many features in ASS scale relative to or are
otherwise affected by the storage resolution of the video file the
subs are bundled with. Adjusting subs for different video resolutions
traditionally took non-trivial effort. Of course this behavior must be
kept by default to preserve compatibility with existing files. But now,
setting both of the new headers to a positive, non-zero value will
replace the video storage resolution with the specified resolution for
the whole file, allowing easy reuse across all encodes of the same
display aspect ratio. This includes PAR effects, so reuse across
anamorphic and non-anamorphic releases is possible too.

Apart from that, some notes for package maintainers:
 * for ASS_FEATURE_WRAP_UNICODE to be usable at runtime, libass must be
   linked against the new optional dependency libunibreak whose source
   currently resides at https://github.com/adah1972/libunibreak
 * switching to a non-recursive build means it is no longer possible to
   invoke make in subdirectories. However this wasn’t really useful
   anyway and everything else, including uninstalled binary locations,
   was kept unchanged.
 * reminder: as announced last release more keys are now authorized for
   signing release tags and tarballs; a signature of any one of those
   keys is sufficient. See the MAINTAINERS file in the last or this
   release’s tarball for a complete list and more details.