Drop unnecessary os2->version != 0xffff check
This kind of check is used internally in FreeType to distinguish a real
OS/2 table from a placeholder that is always present in every FT_Face.
However, the public FT_Get_Sfnt_Table API already performs this check and
returns NULL if a real table is absent: we do not need to recheck this,
and indeed we do not do this anywhere else.