directwrite: add a backup source of name for fallback font
I don't know if it is actually _possible_ in practice for DirectWrite
to return zero WIN32_FAMILY_NAMES for a font, but we do currently guard
against this in ass_directwrite_info_template.h. Use the same logic here.

This name seems somewhat likely to fail to be found by match_fonts
depending on the exact font and version of the Windows environment,
but this is still better than not even trying.