Release 0.17.1
This is an important, bug-fix-only release.

Users of 0.17.0 are advised to upgrade to 0.17.1 soon.
Nothing has changed besides fixing crash and build-stopper bugs,
so there are no side effects.

Users of 0.16.0 are also affected by the buffer overread bug and are
advised to upgrade or at least to backport the relevant fix themselves.

Co-authored-by: Oneric <oneric@oneric.stub>