ass: explicitly read 32bit ints in timestring parts
This matches VSFilter in its traditional ABIs.

It also avoids potential truncation of valid values in ABIs with
smaller ints and signed overflows on ABIs with 64bit ints, like ILP64.
Both are rare in practice.
Still, even as a cosmetic change this makes the supported value range
clearer and matches similar transitions in other parts of the code.