load_file: support unseekable files
Streams do not support seek, thus we are currently unable to load them.
But even for regular files, there are two problems with fseek+ftell.

For one, ftell returns a long, which on Win32’s LLP64 ABI
will artificially limit the loadable file size to 2GiB.

But even if long and size_t are of equal width, ISO C doesn’t require
fseek(.., SEEK_END) to be meaningfully implemented for binary-mode files
as we are using here (and for text-mode ftell need not be meaningful).

Instead just repeatedly load the next chunk of bytes until the file ends
or an error occurs. This avoids those problems, but has the drawback of
likely overallocating the buffer to some degree.
For all current uses of ass_load_file the returned buffer is short-lived
and this is liekly inproblematic, but the following commit will optimise
this for all files whose size is known ahead of time.