Release 0.17.2
This release brings optimized assembly routines for aarch64,
as well as numerous individual improvements and fixes.

Furthermore, Meson/muon is added as a secondary build system restoring
first-party support for MSVC builds. Note however, this build setup is
not at feature parity with autotools and in particular not considered
suitable for packaging purposes on non-Windows platforms.
For more details see our README.