api: constify input parameters where possible
Changed parameters always effectively acted as const
albeit this wasn’t reflected in their declaration yet.
`const char*` is always compatible to `char *` allowing
us to change the following in a backwards-compatible way:
 - filenames
 - codepages (only passed along to iconv which takes const char*)
 - most input data (ecept when passed to iconv as `char** restrict`)

In C++ `const char *const *` is compatible with both
`const char **` and `char **`, but in C it’s only
compatible with the former.
Therefore we cannot change style overrrides for C users,
but we can at least document the effective constness and
make it explicit for C++ users.