ass_utils: add fast compiler-vectorized strchrnul fallback
This mimicks the behavior of fast libc implementations of strchr or strlen:
- Check byte-wise until aligned to vector size
- Check entire vectors at once
- Check byte-wise to determine the exact byte within a vector width

This is implemented using SIMD vectors on x86 and ARM with NEON; on other
platforms, it's implemented using a plain-C technique using scalar registers,
which is comparable to what glibc and musl do.

This may harmlessly overread the end of the array, but can never overread the
page (and thus can't segfault); libc implementations do the same. The function
is tagged with __attribute__((no_sanitize("address"))) accordingly.

This is up to twice as fast as strchr+strlen when strchr returns NULL.
In some cases, it may even be faster than the libc's implementation of
strchrnul (as it uses 16-byte vectorization where the libc may use only
scalar types).