build: add -fno-math-errno to allow inlining of math functions
In some cases, where there is a lot of simple text on the screen, the
ass_render_event and quantize_transform were taking an unexpected amount
of CPU time, causing frame drops.

It turns out that lrint is the root cause of this, as it is not inlined.
Allow it to be lowered to the cvtsd2si instruction.

This change also helps with other similar cases and allows simple math
functions to be inlined.

Overall, rendering performance is 2.2 times faster than before, based on
my test case.