Release 0.17.3
This release fixes a 0.17.2 regression in the fontconfig font provider
as well as some build issues. Furthermore it brings API enhancements
and a performance improvement for a specific case.

The newly added ass_malloc and ass_free API should always be used
when directly interacting with libass-owned or provided data
to be safe on setups with multiple allocators as it’s
not too uncommon on Microsoft Windows.
However libass continues to always use the standard allocators
provided at build time; no currently working setups break.