utils: implement our own lrint() for Visual C++
Commit cd35a4dd added the -fno-math-errno flag which allows lrint() to
be inlined on GCC and Clang.

Unfortunately, Visual C++ can't inline lrint() without -fp:fast. That
flag has some nasty implications, so I think it's better to just write
our own lrint.

This change helps a bit with the video sample from #736. It halves the
number of dropped frames in MPC-HC on my machine (from ~180 to ~80).