meson: set C standard to gnu99 instead of c99
This matches autotools and some of our outline related code
is technically invalid in ISO C99 due to passing a “double[3][3]”
to a function expecting a “const double[3][3]”.
This became officially allowed only in ISO C23 but was supported as a
GNU extensions for much longer. Evidently this is also no build stopper
on MSVC or any other compiler we tested before.

MSVC ofc doesn’t support gnu99, but meson will ignore the option and
in fact already ignored c99 for MSVC only honouring c11, c17 and c18.
https://github.com/mesonbuild/meson/issues/7611