ass_shaper: map Encoding=-2 to FRIBIDI_PAR_RTL
This is useful for forcing RTL rendering of text if they come from
sources that sideload text layout information, such as Youtube where the
text layout is set externally via CSS and isn't present in the sub file
itself. It's not trivial to detect those cases, but it is trivial to
provide a way to force RTL if the user chooses to.

This also implicitly enables WHOLE_TEXT_LAYOUT ASS_Feature